Enhanced check for corruption and moved corruption integration tests in separate tests file
