+++ arctic/store/_ndarray_store.py
@@ -1,5 +1,7 @@
-import logging
 import hashlib
+import logging
+import os
+
 
 from bson.binary import Binary
 import numpy as np
@@ -8,7 +10,7 @@ from pymongo.errors import OperationFailure, DuplicateKeyError
 
 from ..decorators import mongo_retry
 from ..exceptions import UnhandledDtypeException, DataIntegrityException
-from ._version_store_utils import checksum, version_base_or_id
+from ._version_store_utils import checksum, version_base_or_id, _fast_check_corruption
 
 from .._compression import compress_array, decompress
 from six.moves import xrange
@@ -19,7 +21,9 @@ logger = logging.getLogger(__name__)
 _CHUNK_SIZE = 2 * 1024 * 1024 - 2048  # ~2 MB (a bit less for usePowerOf2Sizes)
 _APPEND_SIZE = 1 * 1024 * 1024  # 1MB
 _APPEND_COUNT = 60  # 1 hour of 1 min data
-_CHECK_CORRUPTION_ON_APPEND = False
+
+# Enabling the following has roughly a 5-7% performance hit (off by default)
+_CHECK_CORRUPTION_ON_APPEND = bool(os.environ.get('CHECK_CORRUPTION_ON_APPEND'))
 
 
 def _promote_struct_dtypes(dtype1, dtype2):
@@ -324,21 +328,16 @@ class NdarrayStore(object):
             version['dtype'] = previous_version['dtype']
             version['dtype_metadata'] = previous_version['dtype_metadata']
             version['type'] = self.TYPE
+            
+            # Verify (potential) corruption with append
+            if _CHECK_CORRUPTION_ON_APPEND and _fast_check_corruption(
+                    collection, symbol, previous_version,
+                    check_count=False, check_last_segment=True, check_append_safe=True):
+                logging.warning("Found mismatched segments for {} (version={}). "
+                                "Converting append to concat and rewrite".format(symbol, previous_version['version']))
+                dirty_append = True  # force a concat and re-write (use new base version id)
+
             self._do_append(collection, version, symbol, item, previous_version, dirty_append)
-    
-    def _is_corrupted(self, collection, symbol, version):
-        spec = {'symbol': symbol,
-                'parent': version_base_or_id(version)}
-        # Descending order by creation time, first hit is the last segment belonging to the version.
-        # Sort is trivial as we have up to 60 segments per base version.
-        # If segment counts match, then fetches the latest document, but not the data (small/fast read)
-        curs = collection.find(spec, {'segment': 1}, sort=[('_id', pymongo.DESCENDING)])
-        total_segments = curs.count()
-        if total_segments != version['segment_count']:
-            return True  # don't proceed with fetching from mongo the first hit
-        max_seg = curs.next()
-        max_seg = max_seg['segment'] + 1 if max_seg else 0
-        return max_seg != version['up_to']
 
     def _do_append(self, collection, version, symbol, item, previous_version, dirty_append):
 
@@ -354,14 +353,6 @@ class NdarrayStore(object):
             version['segment_count'] = previous_version['segment_count']
             version['append_count'] = previous_version['append_count']
             version['append_size'] = previous_version['append_size']
-
-        # A flag-controlled check for corrupted symbols.
-        # Note this costs an extra read on the segments collection. Enable it on demand.
-        if _CHECK_CORRUPTION_ON_APPEND and self._is_corrupted(collection, symbol, previous_version):
-            logging.warning(
-                "Found mismatched segments for {} (version={}). Converting append to concat and rewrite".format(
-                symbol, previous_version['version']))
-            dirty_append = True  # force a concat and re-write (use new base version id)
 
         #_CHUNK_SIZE is probably too big if we're only appending single rows of data - perhaps something smaller,
         #or also look at number of appended segments?
